+++ wren-ai-service/src/web/v1/services/__init__.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 from typing import Optional
 
+import orjson
 import pytz
 from pydantic import BaseModel
 
@@ -41,3 +42,16 @@ class Configuration(BaseModel):
     fiscal_year: Optional[FiscalYear] = None
     language: Optional[str] = "English"
     timezone: Optional[Timezone] = Timezone()
+
+
+class SSEEvent(BaseModel):
+    class SSEEventMessage(BaseModel):
+        message: str
+
+        def to_dict(self):
+            return {"message": self.message}
+
+    data: SSEEventMessage
+
+    def serialize(self):
+        return f"data: {orjson.dumps(self.data.to_dict()).decode()}\n\n"
